requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
idna==3.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
vk-api==11.9.9
